pyproject.toml
[project]
name = "entropy-governance"
version = "0.1.0"
description = "Entropy-Governance Duality framework: S∝A (action-governed) vs S∝V (volume-governed), entropic pricing, CREP and Tesseract time-slices."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "GenesisAeon Team", email = "team@genesisaeon.org" }]
requires-python = ">=3.11"
keywords = ["entropy", "governance", "physics", "duality", "sympy", "tesseract"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "sympy>=1.13.0",
    "typer>=0.15.0",
    "pyyaml>=6.0",
    "numpy>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
genesis = ["implosive-genesis>=0.4.0"]
dev = [
    "ruff>=0.6.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
]

[project.scripts]
eg = "entropy_governance.cli:app"

[project.urls]
Repository = "https://github.com/GenesisAeon/entropy-governance"
Issues = "https://github.com/GenesisAeon/entropy-governance/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/entropy_governance"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W", "UP", "C4", "SIM"]
ignore = ["B008"]  # typer uses function calls in defaults

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=entropy_governance --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
